Big update
Fixed issue where program would crash trying to open the main app,
created a new controller class for the main ui (See: MainApp.java)
Added error catching so input for zip can ONLY be a 5-6 long string of
DIGITS.
To do: Make it so you can call the opening of the main app/ui from the
MainApp class, align the strings and images to if statements so they can
be changed, add background image for main app. Fix the text if we don't
like how it looks.