Renamed all the photos to be used in the image
processing classes. The image processing classes

are made so the images are put into an array for

later use. The image processing classes will be

used for when the application runs. If the

weather condition of a certain day corresponds

to a certain photo, you call the proper array

index.